refactor: page header 분리
