+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter as Router } from "react-router-dom"
 import Header from "./widgets/ui/Header.tsx"
 import Footer from "./widgets/ui/Footer.tsx"
-import PostsManagerPage from "./pages/post-manager/PostsManagerPage.tsx"
+import PostsManagerPage from "./pages/post-manager/ui/PostsManagerPage.tsx"
 
 const App = () => {
   return (
